add new hw

+++ src/ru/progwards/java1/lessons/bigints/AbsInteger.java
@@ -2,14 +2,15 @@ package ru.progwards.java1.lessons.bigints;
 
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public class AbsInteger {
 
     public AbsInteger add(AbsInteger num1, AbsInteger num2) {
-        BigDecimal bigNum1 = new BigDecimal(num1.toString());
-        BigDecimal bigNum2 = new BigDecimal(num2.toString());
+        BigInteger bigNum1 = new BigInteger(num1.toString());
+        BigInteger bigNum2 = new BigInteger(num2.toString());
 
-        BigDecimal bigResult = bigNum1.add(bigNum2);
+        BigInteger bigResult = bigNum1.add(bigNum2);
 
         int result = bigResult.intValue();
